move things around
Reduce the amount of noise in the top-level folder. Make room for more
devshell modules.
