+++ devshell/README.md
@@ -1,0 +1,3 @@
+# The devshell CLI
+
+An optional CLI that can be used to make the devshell usage simpler.
