+++ flake.nix
@@ -5,11 +5,11 @@
     let
       # Real developers use Linux for development :-p
       system = "x86_64-linux";
-      pkgs = import ./. { inherit system; };
+      pkgs = import ./nix { inherit system; };
     in
     {
       # Import this overlay into your instance of nixpkgs
-      overlay = import ./overlay.nix;
+      overlay = import ./nix/overlay.nix;
 
       ## Only use for development
       defaultPackage.${system} = pkgs.devshell;
