+++ nix/overlay.nix
@@ -1,5 +1,5 @@
 final: prev:
 {
   devshell = prev.callPackage ./devshell.nix { };
-  mkDevShell = prev.callPackage ./mkDevShell { };
+  mkDevShell = prev.callPackage ./mkDevShell.nix { };
 }
