+++ shell.nix
@@ -1,4 +1,6 @@
 #!/usr/bin/env nix-build
 # Used to test the shell
-{ pkgs ? import ./. { } }:
+{ system ? builtins.currentSystem
+, pkgs ? import ./nix { inherit system; }
+}:
 pkgs.mkDevShell.fromTOML ./devshell.toml
